[eas-cli] Slim supabase/sdk.ts to package constants only
Call shared setupSdkAndConfigAsync with packages/plugin/label at
the connect callsite instead of thin binder wrappers.
